Add and implement nomenclatures Other

+++ Oils/Data/Domains/Product.cs
@@ -18,7 +18,7 @@ namespace Oils.Data.Domains
 
         public int PackageCapacity { get; set; }
 
-        public int PackageWeight { get; set; }
+        public decimal PackageWeight { get; set; }
 
         public ICollection<Lot> Lots { get; set; }
     }
